[BE] Feat: Add getRootFolderList method to FolderController (S12P31A704-200)

See merge request s12-final/S12P31A704!90

+++ be/remembread/src/main/java/com/remembread/card/service/CardSetService.java
@@ -75,16 +75,6 @@ public class CardSetService {
         }
     }
 
-    @Transactional(readOnly = true)
-    public CardSetResponse getCardSetInfo(Long id) {
-        CardSet cardSet = cardSetRepository.getReferenceById(id);
-        CardSetResponse response = new CardSetResponse();
-        List<String> hashtags = hashtagRepository.findAllNamesByCardSetId(cardSet.getId());
-        response.setName(cardSet.getName());
-        response.setHashtags(hashtags);
-        response.setIsPublic(cardSet.getIsPublic());
-        return response;
-    }
     public void forkCardSet(Long cardSetId, Long folderId, Long userId) {
         CardSet cardSet = cardSetRepository.findById(cardSetId).orElseThrow(() -> new GeneralException(ErrorStatus.CARDSET_NOT_FOUND));
         Folder folder = folderRepository.findById(folderId).orElseThrow(() -> new GeneralException(ErrorStatus.FOLDER_NOT_FOUND));
@@ -112,22 +102,44 @@ public class CardSetService {
         cardRepository.saveAll(newCards);
     }
 
+    @Transactional(readOnly = true)
+    public CardSetResponse getCardSetInfo(Long id, User user) {
+        CardSet cardSet = cardSetRepository.findById(id).orElseThrow(() ->
+                new GeneralException(ErrorStatus.CARDSET_NOT_FOUND));
+        if (!cardSet.getIsPublic() && !cardSet.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.CARDSET_NOT_PUBLIC);
+        }
+        CardSetResponse response = new CardSetResponse();
+        List<String> hashtags = hashtagRepository.findAllNamesByCardSetId(cardSet.getId());
+        response.setName(cardSet.getName());
+        response.setHashtags(hashtags);
+        response.setIsPublic(cardSet.getIsPublic());
+        return response;
+    }
 
     @Transactional(readOnly = true)
-    public CardListResponse getCardSetList(Long cardSetId, Integer page, Integer size, String order) {
+    public CardListResponse getCardSetList(Long id, Integer page, Integer size, String order, User user) {
+        CardSet cardSet = cardSetRepository.findById(id).orElseThrow(() ->
+                new GeneralException(ErrorStatus.CARDSET_NOT_FOUND));
+        if (!cardSet.getIsPublic() && !cardSet.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.CARDSET_NOT_PUBLIC);
+        }
         Sort sort = order.equalsIgnoreCase("desc") ?
                 Sort.by("number").descending() :
                 Sort.by("number").ascending();
 
         Pageable pageable = PageRequest.of(page, size, sort);
-        CardSet cardSet = cardSetRepository.getReferenceById(cardSetId);
         List<Card> cards = cardRepository.findAllByCardSet(cardSet, pageable);
         return CardConverter.toCardListResponse(cards);
     }
 
     @Transactional
-    public void updateCardSetInfo(Long cardSetId, CardSetUpdateRequest request) {
-        CardSet cardSet = cardSetRepository.getReferenceById(cardSetId);
+    public void updateCardSetInfo(Long id, CardSetUpdateRequest request, User user) {
+        CardSet cardSet = cardSetRepository.findById(id).orElseThrow(() ->
+                new GeneralException(ErrorStatus.CARDSET_NOT_FOUND));
+        if (!cardSet.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.CARDSET_FORBIDDEN);
+        }
         cardSet.updateName(request.getName());
         cardSet.updateIsPublic(request.getIsPublic());
         this.setHashtag(request.getHashtags(), cardSet);
@@ -135,8 +147,12 @@ public class CardSetService {
     }
 
     @Transactional
-    public void deleteCardSet(Long id) {
-        CardSet cardSet = cardSetRepository.getReferenceById(id);
+    public void deleteCardSet(Long id, User user) {
+        CardSet cardSet = cardSetRepository.findById(id).orElseThrow(() ->
+                new GeneralException(ErrorStatus.CARDSET_NOT_FOUND));
+        if (!cardSet.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.CARDSET_FORBIDDEN);
+        }
         cardSetRepository.delete(cardSet);
     }
 }
